Update to rra-c-util 7.1 and C TAP Harness 4.3
Validate command argument count, the length of command arguments, and
the length of blocks of output from the server fit into the data type
used in the wire protocol.

Check the port argument to remctl and remctld to ensure that it is a
valid port number.

Update to rra-c-util 7.1:

* Avoid spurious test failures from the network library.
* Fix configure output when a Kerberos install prefix was provided.
* Fix new warnings in GCC 7 and add new warning flags.
* Fix all warnings from the Clang static analyzer.
* Fix warnings under Clang with most warnings enabled.

Update to C TAP Harness 4.3:

* On test failures, report left/right instead of wanted/expected.
* Fix string comparisons involving NULL pointers.

+++ server/internal.h
@@ -2,9 +2,9 @@
  * Internal support functions for the remctld daemon.
  *
  * Written by Russ Allbery <eagle@eyrie.org>
+ * Copyright 2015-2016, 2018 Russ Allbery <eagle@eyrie.org>
  * Copyright 2016 Dropbox, Inc.
- * Copyright 2015, 2016 Russ Allbery <eagle@eyrie.org>
- * Copyright 2006, 2007, 2008, 2009, 2010, 2012, 2014
+ * Copyright 2006-2010, 2012, 2014
  *     The Board of Trustees of the Leland Stanford Junior University
  *
  * See LICENSE for licensing terms.
@@ -81,7 +81,7 @@ struct client {
 /* Holds the configuration for a single command. */
 struct rule {
     char *file;                 /* Config file name. */
-    int lineno;                 /* Config file line number. */
+    size_t lineno;              /* Config file line number. */
     struct vector *line;        /* The split configuration line. */
     char *command;              /* Command (first argument). */
     char *subcommand;           /* Subcommand (second argument). */
